Delete index.html
I have deployed my web application from the main file and root folder.  But my index file is in a subfolder and I can't figure out how to move it.  So I uploaded a new index file from my local computer into the root directory which I thought was identical but it was an earlier version.  I am deleting this in the hope I can figure out how simply to move the index file.